Load clang-format config directly from original file
Instead of copying the `.clang-format` file to each package, load the
config from the original file and pass it as a YAML string to the
`clang-format` command.
Add a simply Python script `yaml2oneline` for this, which loads a YAML
file and prints it as a one-line string ("{key1: value1, ...}").  The
config is loaded into a cmake variable using that script and then passed
on to `clang-format`.

Further do some refactoring:
- Move the executable scripts from `resources` to `scripts`.
- Rename the `FORMAT_CODE_DURING_BUILD` to the shorter `FORMAT_CODE`.