Fail formatting target if there is any output
clang-format may not return a non-zero exit code in all error cases
(e.g. older versions just fall back to a default style if there is some
syntax error in the provided style).  Assuming that there is normally no
output to stdout, simply check the output of the call and consider it
failed if it is non-empty.