Define package path variables and export them via CFG_EXTRAS
Define path variables for the package root as well as the relevant
subdirectories.
Add a template file in which these paths will be stored and which is
exported via CFG_EXTRAS.  This way, the path variables can be used in
the other cmake files without the need to include anything.