clang-format: Preserve include blocks (#35)
Explicitly tell clang-format to preserve include blocks when sorting
includes.  For example

    #include <a.h>
    #include <c.h>

    #include <b.h>

Is not changed.  Without this setting the two blocks are merged into one
in newer clang versions:

    #include <a.h>
    #include <b.h>
    #include <c.h>

Note: It seems that the default behaviour of clang-format changed.  It
was okay without this using version 6 but is needed with version 9.