Do not separately search for Python executable
The Python executable is already search for by `FindPython.cmake`, which
respects `PYTHON_VERSION` and defaults to Python 3.

This separate search for the executable is run before `FindPython.cmake`
(thus has precedence) and defaults to `python`, which is typically
Python 2 on a standard Ubuntu installation.
The result can be a broken build as some parts of the workspace are
installed for Python 2 and others for Python 3.

It seems that just removing the search for the executable here is okay,
as it is then searched for later on by `FindPython.cmake`.