Use install(PROGRAMS ...) for scripts
The previous command only set the executable flag for the owner of the
scripts, which is a problem if it is supposed to be used by other users
as well (e.g. in a container that is build by root but then used by
"normal" users).

Instead use `install(PROGRAMS ...)` which takes care of the permissions.
With this, the installed scripts have permissions `-rwxr-xr-x`, allowing
all users to execute them.