Replace mpi_cpp_format
Instead of our own implementation use the `run-clang-format` script from
[here](https://github.com/Sarcasm/run-clang-format).

The main advantage is that it allows specifying files/directories that
should be excluded from the formatting, either via `--exclude` or by
adding a file `.clang-format-ignore` to the repository.

The copy of run-clang-format added here is extended a bit to allow
specifying a _clang-format file.

The `cpp_format.py` module is removed and `mpi_cpp_format` is replaced
by a thin wrapper around `run-clang-format` that sets some parameters
such that the behaviour is mostly the same as with the old
`mpi_cpp_format`.