Update list of default EXCLUDE_PATTERNS
Add `src/` and `test/` to excluded directories.

Prepend all exclude patterns with the project source dir to avoid
matches in the higher level part of the path (e.g. adding `*/src/*` will
exclude everything because in a colcon workspace all packages are in a
directory called `src/`...).

Remove some more unusual patterns (build and resources).  For
corner-cases where one of them would still be needed, there is now the
option to add custom excludes in the `add_documentation` macro.

This should also fix a critical issue that C++ API documentation is not
build on Bamboo.  I believe this is because on Bamboo the workspace is
build in a directory like
`/home/bambooagent/bamboo-agent-home/xml-data/build-dir/MGC-SHAREDM-U2/`
and since the unrestricted `*/*build*/*` was matching the
`build-dir`-part, _all_ C++ files of the whole workspace got excluded.